backend/requirements.txt
fastapi
uvicorn
mcp[fastapi]
pyyaml
pydantic
python-dotenv
chromadb
jinja2
python-multipart
